refactor: Remove legacy journey node mappers and migrate to new IJourneyNodeMapper approach
- Deleted outdated `JourneyNodeMapper` interface and its implementations for all journey node types.
- Introduced new `IJourneyNodeMapper` interface with enhanced flexibility and support for dependency injection.
- Reimplemented node-specific mappers using the updated interface: triggers, conditions, actions, and extractors.
- Streamlined serialization logic and node mapping resolution process with `resolveNode` for next/notMatch references.

+++ src/main/kotlin/infrastructure/journey/condition/ConditionJourneyNodeNomenclature.kt
@@ -53,20 +53,20 @@ object ConditionJourneyNodeNomenclature : JourneyNodeNomenclature<IConditionJour
             )
             "numberInRange" -> NumberInRangeConditionJourneyNode(
                 inputKey = inputKey,
-                min = map["min"] as Number,
-                max = map["max"] as Number,
+                min = (map["min"] as Number).toDouble(),
+                max = (map["max"] as Number).toDouble(),
             )
             "numberMoreThan" -> NumberMoreThanConditionJourneyNode(
                 inputKey = inputKey,
-                threshold = map["threshold"] as Number,
+                threshold = (map["threshold"] as Number).toDouble(),
             )
             "numberLessThan" -> NumberLessThanConditionJourneyNode(
                 inputKey = inputKey,
-                threshold = map["threshold"] as Number,
+                threshold = (map["threshold"] as Number).toDouble(),
             )
             "numberEqual" -> NumberEqualConditionJourneyNode(
                 inputKey = inputKey,
-                target = map["target"] as Number,
+                target = (map["target"] as Number).toDouble(),
             )
             "stringEqual" -> StringEqualConditionJourneyNode(
                 inputKey = inputKey,
